+++ client/my-sites/subscribers/components/subscriber-data-views/style.scss
@@ -2,6 +2,7 @@
 @import '@wordpress/base-styles/breakpoints';
 @import '@wordpress/base-styles/mixins';
 @import '@wordpress/theme/design-tokens.css';
+@import '@wordpress/admin-ui/build-style/style.css';
 @import '@wordpress/dataviews/build-style/style.css';
 
 $padding-standard: 16px;
