--- client/my-sites/subscribers/components/subscribers-header/index.ts
@@ -1,1 +1,0 @@
-export { SubscribersHeader } from './subscribers-header';
